Split Executor and Factory modules
refactor(executor): extract Executor, Processor, InfraFunction,
 Plant, InfraStructure, and TF helpers into modules;
refactor(factory): Factory.accept/assemble/produce manufacturing
 cell with stable package re-exports
