+++ cats/executor/__init__.py
@@ -1,86 +1,24 @@
-from cats.executor.function import InfraFunction, Processor
+"""Ephemeral Executor of the Architectural Quantum and its halves."""
+from cats.executor.executor import Executor
+from cats.executor.function import Function, InfraFunction, Processor
 from cats.executor.structure import (
     InfraStructure,
     Plant,
+    Structure,
+    modules_installed,
     read_applied_structure_cid,
     write_applied_structure_cid,
 )
 
-
-class Structure:
-    def __init__(self, service, structure_cid):
-        self.service = service
-        self.bom_json_cid = self.service.bom_json_cid
-        self.infraStructure: InfraStructure = InfraStructure(service=self.service, structure_cid=structure_cid)
-        self.plant: Plant = self.infraStructure.compose()
-
-    def redeploy(self):
-        print()
-        print()
-        print('Re-Deploy Structure!')
-        # `destroy` needs providers already installed to even load their
-        # schemas, same as `apply`/`plan` - so `initialize` must run first,
-        # not just before `apply` below.
-        self.infraStructure.initialize()
-        self.infraStructure.destroy()
-        self.infraStructure.apply()
-        self.plant.rebuilt = True
-
-    def deploy(self):
-        print()
-        print()
-        print('Deploy Structure!')
-        self.infraStructure.initialize()
-        self.infraStructure.apply()
-        self.plant.rebuilt = False
-
-    def reconcile(self):
-        """Materialize this CAT's Structure, skipping the destructive
-        rebuild when the incoming (content-addressed) structure_cid
-        matches what's already applied.
-
-        Structure is content-addressed like everything else in CATs.
-        Destroying and rebuilding the Plant (kind cluster + Helm
-        releases) for a Structure whose content hasn't changed is the
-        same redundant recomputation content-addressing is meant to
-        avoid elsewhere in CATs; `apply()` alone is Terraform's own
-        declarative reconciliation, and it's a fast no-op when nothing
-        changed.
-
-        Returns a snapshot of the resulting Plant (see `Plant.snapshot()`),
-        so callers can record what this Structure actually produced
-        alongside Function's output in the CAT's BOM.
-        """
-        structure_cid = self.infraStructure.structure_cid
-        structure_home = self.infraStructure.INPUT_STRUCTURE_HOME
-        applied_cid = read_applied_structure_cid(structure_home)
-        if structure_cid and applied_cid == structure_cid:
-            print(f'Structure {structure_cid} already applied; reconciling in place.')
-            self.deploy()
-        else:
-            self.redeploy()
-        if structure_cid:
-            write_applied_structure_cid(structure_home, structure_cid)
-        return self.plant.snapshot()
-
-
-class Function:
-    def __init__(self, service, function_cid):
-        self.service = service
-        self.CAT_HOME = None
-        self.infraFunction: InfraFunction = InfraFunction(service=self.service, function_cid=function_cid)
-        self.processor: Processor = self.infraFunction.compose()
-        self.ingress_data_cid = None
-        self.integration_data_cid = None
-        self.egress_data_cid = None
-        self.invoice_data_cid = None
-
-    def execute(self, object_store, plant, transport):
-        self.ingress_data_cid, self.integration_data_cid, self.egress_data_cid = \
-            self.processor.process(
-                object_store=object_store,
-                plant=plant,
-                transport=transport,
-            )
-        self.invoice_data_cid = self.processor.invoice_data_cid
-        return self.ingress_data_cid, self.integration_data_cid, self.egress_data_cid
+__all__ = [
+    'Executor',
+    'Function',
+    'InfraFunction',
+    'Processor',
+    'Structure',
+    'InfraStructure',
+    'Plant',
+    'modules_installed',
+    'read_applied_structure_cid',
+    'write_applied_structure_cid',
+]
